Clean up a variety of issues with spm, ps/pss, sdb, ssc, hpe

+++ components/ysw_sdb/include/ysw_sdb.h
@@ -13,7 +13,7 @@
 #include "lvgl.h"
 
 typedef void (*ysw_sdb_string_cb_t)(void *context, const char *new_value);
-typedef void (*ysw_sdb_choice_cb_t)(void *context, uint8_t new_index);
+typedef void (*ysw_sdb_choice_cb_t)(void *context, uint16_t new_index);
 typedef void (*ysw_sdb_switch_cb_t)(void *context, bool new_value);
 typedef void (*ysw_sdb_checkbox_cb_t)(void *context, bool new_value);
 typedef void (*ysw_sdb_button_cb_t)(void *context);
@@ -33,7 +33,7 @@ ysw_sdb_t *ysw_sdb_create_standard(lv_obj_t *parent, const char *title, void *co
 ysw_sdb_t* ysw_sdb_create_custom(lv_obj_t *parent, const char *title, const ysw_ui_btn_def_t buttons[], void *context);
 lv_obj_t *ysw_sdb_add_separator(ysw_sdb_t *sdb, const char *name);
 lv_obj_t *ysw_sdb_add_string(ysw_sdb_t *sdb, const char *name, const char *value, void *cb);
-lv_obj_t *ysw_sdb_add_choice(ysw_sdb_t *sdb, const char *name, uint8_t value, const char *options, void *cb);
+lv_obj_t *ysw_sdb_add_choice(ysw_sdb_t *sdb, const char *name, uint16_t value, const char *options, void *cb);
 lv_obj_t *ysw_sdb_add_switch(ysw_sdb_t *sdb, const char *name, bool value, void *cb);
 lv_obj_t *ysw_sdb_add_checkbox(ysw_sdb_t *sdb, const char *name, const char *text, bool value, void *cb);
 lv_obj_t *ysw_sdb_add_button(ysw_sdb_t *sdb, const char *name, void *cb);
